fix(files): use explicit archive extraction route
