+++ apps/sim/app/api/workspaces/[id]/files/[fileId]/route.test.ts
@@ -6,7 +6,6 @@ import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 const mocks = vi.hoisted(() => ({
   getSession: vi.fn(),
-  extract: vi.fn(),
   rename: vi.fn(),
   deleteItems: vi.fn(),
   getUserEntityPermissions: vi.fn(),
@@ -14,13 +13,6 @@ const mocks = vi.hoisted(() => ({
 }))
 
 vi.mock('@/lib/auth', () => ({ getSession: mocks.getSession }))
-
-vi.mock('@/lib/workspace-files/application/extract-workspace-file', () => ({
-  extractWorkspaceFile: {
-    operation: { id: 'files.extract_archive', minimumRole: 'write', workspaceApiKey: 'deny' },
-    execute: mocks.extract,
-  },
-}))
 
 vi.mock('@/lib/workspace-files/application/rename-workspace-file', () => ({
   renameWorkspaceFile: {
@@ -46,8 +38,7 @@ import {
   WorkspaceApiKeyScopeAuthorizationError,
 } from '@/lib/core/application'
 import { OrchestrationError } from '@/lib/core/orchestration/types'
-import { ArchiveError } from '@/lib/uploads/archive'
-import { PATCH, POST } from '@/app/api/workspaces/[id]/files/[fileId]/route'
+import { PATCH } from '@/app/api/workspaces/[id]/files/[fileId]/route'
 
 const WORKSPACE_ID = 'workspace-1'
 const FILE_ID = 'wf_1'
@@ -59,15 +50,6 @@ function callRename(body: unknown) {
       method: 'PATCH',
       headers: { 'Content-Type': 'application/json' },
       body: JSON.stringify(body),
-    }),
-    context
-  )
-}
-
-function callExtract() {
-  return POST(
-    new NextRequest(`http://localhost:3000/api/workspaces/${WORKSPACE_ID}/files/${FILE_ID}`, {
-      method: 'POST',
     }),
     context
   )
@@ -97,7 +79,6 @@ describe('PATCH /api/workspaces/[id]/files/[fileId]', () => {
       session: { id: 'session-1' },
     })
     mocks.rename.mockResolvedValue({ file: fileRecord() })
-    mocks.extract.mockResolvedValue({ folderName: 'bundle', extractedCount: 2, skippedCount: 0 })
   })
 
   it('authenticates before parsing the request', async () => {
@@ -187,51 +168,5 @@ describe('PATCH /api/workspaces/[id]/files/[fileId]', () => {
     expect(await response.json()).toEqual({
       error: 'Internal server error',
     })
-  })
-})
-
-describe('POST /api/workspaces/[id]/files/[fileId]', () => {
-  beforeEach(() => {
-    vi.clearAllMocks()
-    mocks.getSession.mockResolvedValue({
-      user: { id: 'user-1' },
-      session: { id: 'session-1' },
-    })
-    mocks.extract.mockResolvedValue({ folderName: 'bundle', extractedCount: 2, skippedCount: 0 })
-  })
-
-  it('passes a session principal and canonical assertion to the extraction use case', async () => {
-    const response = await callExtract()
-
-    expect(response.status).toBe(200)
-    expect(await response.json()).toEqual({
-      success: true,
-      folderName: 'bundle',
-      extractedCount: 2,
-      skippedCount: 0,
-    })
-    expect(mocks.extract).toHaveBeenCalledWith({
-      principal: { kind: 'session', userId: 'user-1', sessionId: 'session-1' },
-      input: { fileId: FILE_ID, assertedWorkspaceId: WORKSPACE_ID },
-      request: expect.anything(),
-    })
-  })
-
-  it('authenticates before invoking extraction', async () => {
-    mocks.getSession.mockResolvedValue(null)
-
-    const response = await callExtract()
-
-    expect(response.status).toBe(401)
-    expect(mocks.extract).not.toHaveBeenCalled()
-  })
-
-  it('returns a caller-safe error for an invalid zip', async () => {
-    mocks.extract.mockRejectedValue(new ArchiveError('invalid', 'Not a valid .zip archive.'))
-
-    const response = await callExtract()
-
-    expect(response.status).toBe(400)
-    expect(await response.json()).toEqual({ error: 'Not a valid .zip archive.' })
   })
 })
